# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar GreenPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.stellargreenpay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PublicKeyParam:
      name: publicKey
      in: path
      required: true
      description: Stellar public key.
      schema:
        type: string
        pattern: '^G[A-Z0-9]{55}$'
...
    JobId:
      name: id
      in: path
      required: true
      description: Escrow job identifier.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    Profile:
      type: object
      properties:
        publicKey:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        totalDonatedXLM:
          type: string
        projectsSupported:
          type: integer
        badges:
          type: array
          items:
            type: object
            properties:
              tier:
                type: string
              earnedAt:
                type: string
                format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        clientPublicKey:
          type: string
        freelancerPublicKey:
          type: string
        amountEscrowXlm:
          type: string
        status:
          type: string
        releaseTransactionHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProfileRequest:
      type: object
      required: [publicKey]
      properties:
        publicKey:
          type: string
          pattern: '^G[A-Z0-9]{55}$'
        displayName:
          type: string
          maxLength: 30
        bio:
          type: string
          maxLength: 300
...
paths:
...
  /api/profiles/{publicKey}:
    get:
      tags: [Profiles]
      summary: Get a profile by public key
      parameters:
        - $ref: '#/components/parameters/PublicKeyParam'
      responses:
        '200':
          description: Profile data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/profiles:
    post:
      tags: [Profiles]
      summary: Create or update a profile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '200':
          description: Profile upserted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/jobs/{id}/release:
    patch:
      tags: [Jobs]
      summary: Release escrow for a job
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [releaseTransactionHash]
              properties:
                releaseTransactionHash:
                  type: string
                  pattern: '^[a-fA-F0-9]{64}$'
      responses:
        '200':
          description: Job updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get a job by id
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
...